requirements.txt
charm-crypto     # For pairing-based cryptography and ECC
cryptography    # For hashes and signatures
asyncio                 # Built-in, no need to install separately

# To install above dependencies:
# pip install -r requirements.txt